minor refactor around the metrics repos
